+++ tests/test_dcos_e2e/backends/docker/test_distributions.py
@@ -59,6 +59,7 @@ def _oss_distribution_test(
     ) as cluster:
         cluster.install_dcos_from_path(
             build_artifact=oss_artifact,
+            dcos_config=cluster.base_config,
             log_output_live=True,
         )
         cluster.wait_for_dcos_oss()
@@ -97,7 +98,10 @@ def _enterprise_distribution_test(
     ) as cluster:
         cluster.install_dcos_from_path(
             build_artifact=enterprise_artifact,
-            extra_config=config,
+            dcos_config={
+                **cluster.base_config,
+                **config,
+            },
             log_output_live=True,
         )
         cluster.wait_for_dcos_ee(