+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Next
 - Show a formatted error rather than a traceback if Docker cannot be connected to.
 - Custom backends' must now implement a ``base_config`` method.
 - Custom backends' installation methods must now take ``dcos_config`` rather than ``extra_config``.
+- ``Cluster.install_dcos_from_url`` and ``Cluster.install_dcos_from_path`` now take ``dcos_config`` rather than ``extra_config``.
 
 2018.05.17.0
 ------------